ov_init predeclares
