+++ ov_init/src/dynamic/DynamicInitializer.cpp
@@ -21,15 +21,29 @@
 
 #include "DynamicInitializer.h"
 
+#include "ceres/Factor_GenericPrior.h"
+#include "ceres/Factor_ImageReprojCalib.h"
+#include "ceres/Factor_ImuCPIv1.h"
+#include "ceres/State_JPLQuatLocal.h"
+#include "utils/helper.h"
+
+#include "cpi/CpiV1.h"
+#include "feat/Feature.h"
+#include "feat/FeatureDatabase.h"
+#include "types/IMU.h"
+#include "types/Landmark.h"
+#include "utils/colors.h"
+#include "utils/print.h"
+#include "utils/quat_ops.h"
+#include "utils/sensor_data.h"
+
 using namespace ov_core;
 using namespace ov_type;
 using namespace ov_init;
 
 bool DynamicInitializer::initialize(double &timestamp, Eigen::MatrixXd &covariance, std::vector<std::shared_ptr<ov_type::Type>> &order,
                                     std::shared_ptr<ov_type::IMU> &_imu, std::map<double, std::shared_ptr<ov_type::PoseJPL>> &_clones_IMU,
-                                    std::unordered_map<size_t, std::shared_ptr<ov_type::Landmark>> &_features_SLAM,
-                                    std::unordered_map<size_t, std::shared_ptr<ov_type::PoseJPL>> &_calib_IMUtoCAM,
-                                    std::unordered_map<size_t, std::shared_ptr<ov_type::Vec>> &_cam_intrinsics) {
+                                    std::unordered_map<size_t, std::shared_ptr<ov_type::Landmark>> &_features_SLAM) {
 
   // Get the newest and oldest timestamps we will try to initialize between!
   auto rT1 = boost::posix_time::microsec_clock::local_time();
